Save modified entities.

+++ src/storage.rs
@@ -18,8 +18,8 @@ pub trait EntityStorageFactory {
 #[derive(Debug)]
 pub struct PersistedEntity {
     pub key: String,
-    pub gid: u32,
-    pub version: u32,
+    pub gid: u64,
+    pub version: u64,
     pub serialized: String,
 }
